E_CLASSROOM-306: Created another that will render the dropdown component
